Generalize InfoTopic component
