+++ res/css/_components.pcss
@@ -170,6 +170,7 @@
 @import "./views/elements/_EditableItemList.pcss";
 @import "./views/elements/_ErrorBoundary.pcss";
 @import "./views/elements/_EventTilePreview.pcss";
+@import "./views/elements/_ExpandableBox.pcss";
 @import "./views/elements/_ExternalLink.pcss";
 @import "./views/elements/_FacePile.pcss";
 @import "./views/elements/_Field.pcss";
