+++ res/css/views/elements/_ExpandableBox.pcss
@@ -1,0 +1,38 @@
+/*
+Copyright 2022 The Matrix.org Foundation C.I.C.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+.mx_RoomSummaryCard_infoTopic {
+    .mx_RoomSummaryCard_infoTopic_text {
+        color: $secondary-content;
+        transition: height 0.5 ease-in-out;
+
+        overflow: hidden;
+        -webkit-box-orient: vertical;
+        display: -webkit-box;
+
+        /* Properties below are controlled by the respective react component */
+        /* line-height: $lineHeight; */
+        /* -webkit-line-clamp: $lines; */ /* See: https://drafts.csswg.org/css-overflow-3/#webkit-line-clamp */
+    }
+
+    .mx_RoomSummaryCard_infoTopic_toggle {
+        float: right;
+        border: none;
+        color: $accent;
+        background: none;
+        cursor: pointer;
+    }
+}
