+++ res/css/views/right_panel/_RoomSummaryCard.pcss
@@ -89,29 +89,6 @@ limitations under the License.
         }
     }
 
-    .mx_RoomSummaryCard_infoTopic {
-        .mx_RoomSummaryCard_infoTopic_text {
-            color: $secondary-content;
-            transition: height 0.5 ease-in-out;
-
-            overflow: hidden;
-            -webkit-box-orient: vertical;
-            display: -webkit-box;
-
-            /* Properties below are controlled by the respective react component */
-            /* line-height: $lineHeight; */
-            /* -webkit-line-clamp: $lines; */ /* See: https://drafts.csswg.org/css-overflow-3/#webkit-line-clamp */
-        }
-
-        .mx_RoomSummaryCard_infoTopic_toggle {
-            float: right;
-            border: none;
-            color: $accent;
-            background: none;
-            cursor: pointer;
-        }
-    }
-
     .mx_RoomSummaryCard_aboutGroup {
         .mx_RoomSummaryCard_Button {
             padding-left: 44px;
